requirements.txt
Flask
torch
timm
Pillow
torchvision